% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff.R
\name{additions}
\alias{additions}
\title{Compute a patch of additions on a recursive object.}
\usage{
additions(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Compute a patch of additions on a recursive object.
}

