% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/patch.R
\name{atomic_differences_patch}
\alias{atomic_differences_patch}
\title{Generate a patch for two atomic objects that are close in values.}
\usage{
atomic_differences_patch(old_object, new_object, transition = TRUE)
}
\arguments{
\item{old_object}{atomic.}

\item{new_object}{atomic.}

\item{transition}{logical. Whether or not to use a transition depending
on how many element do not match. Namely, if over 50% do not match in
from a random sample of 100 elements (so most of \code{new_object} is
probably different than \code{old_object}) then replace it completely
with a trivial patch; otherwise, perform a more subtle calculation
using \code{base::!=} and stores only exactly which elements changed.}
}
\description{
Generate a patch for two atomic objects that are close in values.
}
\examples{
x <- 1:10; y <- x; y[1] <- 5
patch <- objectdiff:::atomic_differences_patch(x, y)
stopifnot(identical(y, patch(x)))
}

