% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/patch.R
\name{attributes_patch}
\alias{attributes_patch}
\title{Assume two objects are identical and only patch their attributes.}
\usage{
attributes_patch(old_object, new_object)
}
\arguments{
\item{old_object}{atomic.}

\item{new_object}{atomic.}
}
\description{
Assume two objects are identical and only patch their attributes.
}

