% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{commit<-.tracked_environment}
\alias{commit<-.tracked_environment}
\title{Commit a change to a tracked environment.}
\usage{
\method{commit}{tracked_environment}(env) <- value
}
\arguments{
\item{env}{tracked_environment.}

\item{value}{character. Commit message. May be \code{NULL}.}
}
\description{
Committing a change is equivalent to storing a patch object (see
\code{\link{objectdiff}}). For the moment, any comment will produce
a single patch by "squishing" any changes made to the environment.
If you are iterating over an environment object many many times,
you should probably batch changes if it lives in a tracked environment.
Otherwise, for example, making 1,000,000 small modifications to
a data.frame before commiting could cause a long time to recompute
the "final patch".
}
\seealso{
\code{\link{objectdiff}})
}

