% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracked_environment.R
\name{commit<-.tracked_environment}
\alias{commit}
\alias{commit<-.tracked_environment}
\title{Commit a change to a tracked environment.}
\usage{
\method{commit}{tracked_environment}(env) <- value

commit(env, value = NULL)
}
\arguments{
\item{env}{tracked_environment.}

\item{value}{character. Commit message. May be \code{NULL}.}
}
\description{
Committing a change is equivalent to storing a patch object (see
\code{\link{objectdiff}}). For the moment, any comment will produce
a single patch by "squishing" any changes made to the environment.
If you are iterating over an environment object many many times,
you should probably batch changes if it lives in a tracked environment.
Otherwise, for example, making 1,000,000 small modifications to
a data.frame before commiting could cause a long time to recompute
the "final patch".
}
\examples{
x <- tracked_environment()
x$foo <- 1
commit(x) <- 'First message'
x$foo <- 2
}
\seealso{
\code{\link{objectdiff}})
}

