% Generated by roxygen2 (4.0.2): do not edit by hand
\name{commit.tracked_environment}
\alias{commit.tracked_environment}
\title{Commit a change to a tracked environment.}
\usage{
\method{commit}{tracked_environment}(env, value)
}
\arguments{
\item{env}{tracked_environment.}

\item{value}{character. Commit message. May be \code{NULL}.}
}
\description{
Committing a change is equivalent to storing a patch object (see
\code{\link{objectdiff}})
}
\seealso{
\code{\link{objectdiff}})
}

