% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/tracked_environment.R
\name{commits}
\alias{commits}
\title{A named list of all commits.}
\usage{
commits(env)
}
\arguments{
\item{env}{tracked_environment.}
}
\value{
a named list of all commits, with each name corresponding
  to the commit message, and each value a \code{patch} object
  (a function that you can apply to a tracked environment to
  perform that commit).
}
\description{
A named list of all commits.
}

