% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/diff.R
\name{diff}
\alias{diff}
\title{Compute a patch that takes a diff of a recursive object (list or tracked_environment).}
\usage{
diff(old_object, new_object)
}
\arguments{
\item{old_object}{list or tracked_environment.}

\item{new_object}{list or tracked_environment.}
}
\description{
Compute a patch that takes a diff of a recursive object (list or tracked_environment).
}

