% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/patch.R
\name{differences_patch}
\alias{differences_patch}
\title{Generate a patch given a recording of differences between two objects.}
\usage{
differences_patch(old_object, new_object, differences)
}
\arguments{
\item{old_object}{atomic.}

\item{new_object}{atomic.}

\item{differences}{logical. The differences in first and second object.
These should be calculated externally because a different approach
could be used for different objects (e.g., lists versus atomic;
in the former we would need \code{base::identical} on each element,
whereas in the latter we could use \code{base::`!=`}).}
}
\description{
This patch will use another patch to record changes to attributes
(if any). Otherwise, given indices of changed object, it will
generate a patch over those indices.
}

