% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/tracked_environment.R
\name{force_push}
\alias{force_push}
\title{Force push a tracked environment to a given commit.}
\usage{
force_push(env, commit)
}
\arguments{
\item{env}{tracked_environment.}

\item{commit}{integer or character. If character, the commit with this
name will be attempted for the force push. If there are multiple commits
with this same name, a warning will be issued.}
}
\description{
Force push a tracked environment to a given commit.
}
\examples{
env <- tracked_environment()
env$x <- 1
commit(env) <- 'first commit'
env$y <- 2
commit(env) <- 'second commit'
force_push(env, 'first commit') # equivalent to force_push(env, 1)
stopifnot(identical(as.list(environment(env)), list(x = 1)))
}

