% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{heterogeneous_list_patch}
\alias{heterogeneous_list_patch}
\title{Compute a patch between two lists of different length or different names.}
\usage{
heterogeneous_list_patch(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\value{
a \code{\link{patch}} translating the \code{old_object} to
   the \code{new_object}.
}
\description{
Compute a patch between two lists of different length or different names.
}

