% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{is.identity_patch}
\alias{is.identity_patch}
\title{Check if an R object is an identity patch.}
\usage{
is.identity_patch(fn)
}
\arguments{
\item{fn}{function. Any function.}
}
\value{
TRUE or FALSE according as the function is or is not an identity patch.
}
\description{
The identity patch just returns the object that was passed in.
}

