% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{modifications}
\alias{modifications}
\title{Compute a patch of modifications on a recursive object.}
\usage{
modifications(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Compute a patch of modifications on a recursive object.
}

