% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.frame.R
\docType{methods}
\name{objectdiff,data.frame,data.frame-method}
\alias{objectdiff,data.frame,data.frame-method}
\title{Record what changes have been made to a tracked environment.}
\usage{
\S4method{objectdiff}{data.frame,data.frame}(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Record what changes have been made to a tracked environment.
}

