% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{objectdiff,environment,environment-method}
\alias{objectdiff,environment,environment-method}
\title{Transform an environment into another environment.}
\usage{
\S4method{objectdiff}{environment,environment}(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Transform an environment into another environment.
}

