% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{objectdiff}
\alias{objectdiff}
\alias{objectdiff-package}
\title{Objectdiff is an R package for measuring differences
between arbitrary R objects using patches, a la Git. For example,
if you have slightly modified a data.frame, and you would like to
record the change, it would be more efficient to only record what
is different, rather than a full copy of both data sets.}
\description{
Objectdiff is an R package for measuring differences
between arbitrary R objects using patches, a la Git. For example,
if you have slightly modified a data.frame, and you would like to
record the change, it would be more efficient to only record what
is different, rather than a full copy of both data sets.
}

