% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectdiff.R
\docType{class}
\name{patch-class}
\alias{patch-class}
\title{A patching function to transform one R object into another.}
\description{
The base class for all results of the \code{objectdiff} function.
For example, \code{objectdiff(data1, data2)} will return a function that
converts \code{data1} into \code{data2}, and that function will be of class
\code{patch}.
}

