% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{reorder_names_patch}
\alias{reorder_names_patch}
\title{Re-order names according to the names of object.}
\usage{
reorder_names_patch(object)
}
\arguments{
\item{object}{ANY. Any named R object (e.g., a named list).}
}
\description{
Re-order names according to the names of object.
}
\note{
This only works on uniquely named objects.
}

