% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tracked_environment.R
\name{rollback<-.tracked_environment}
\alias{rollback<-.tracked_environment}
\title{Roll back commits to an earlier version of the tracked environment.}
\usage{
\method{rollback}{tracked_environment}(env) <- value
}
\arguments{
\item{env}{tracked_environment.}

\item{value}{integer. Number of commits to roll back.}
}
\description{
Roll back commits to an earlier version of the tracked environment.
}

