% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/squish.R
\name{squish_patches}
\alias{squish_patches}
\title{Squish a sequence of patches off a base object into one.}
\usage{
squish_patches(base, patches)
}
\arguments{
\item{base}{ANY. R object to base the squishing off.}

\item{patches}{list. A list of patches (functions derived from
\code{\link{objectdiff}}).}
}
\value{
a single \code{patch}.
}
\description{
Squish a sequence of patches off a base object into one.
}
\examples{
iris2 <- iris; iris2[1,1] <- 1
patch1 <- objectdiff(iris, iris2)
iris3 <- iris2; iris3[1,1] <- 2
patch2 <- objectdiff(iris2, iris3)
patch3 <- squish_patches(iris, list(patch1, patch2))
stopifnot(identical(iris3, patch3(iris)))
}
\seealso{
\code{\link{objectdiff}}
}

