% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_env}
\alias{copy_env}
\title{Copy one environment into another recursively.}
\usage{
copy_env(to, from)
}
\arguments{
\item{to}{environment. The new environment.}

\item{from}{environment. The old environment.}
}
\description{
Copy one environment into another recursively.
}

