% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.R
\name{estimate_size}
\alias{estimate_size}
\title{Estimate the size of a list stochastically.}
\usage{
estimate_size(list, sampling_percentage = 0.05)
}
\arguments{
\item{list}{The list object whose size we are estimating.}

\item{sampling_percentage}{numeric. Default is \code{0.05}.}
}
\description{
We assume each element in the list is approximately the same size,
sample some small percentage, and multiply by that amount.
}

