% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\docType{methods}
\name{objectdiff,list,list-method}
\alias{objectdiff,list,list-method}
\title{Transform a list into another list.}
\usage{
\S4method{objectdiff}{list,list}(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Transform a list into another list.
}

