% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic.R
\docType{methods}
\name{objectdiff,numeric,numeric-method}
\alias{objectdiff,numeric,numeric-method}
\title{Transform an atomic vector into an atomic vector.}
\usage{
\S4method{objectdiff}{numeric,numeric}(old_object, new_object)
}
\arguments{
\item{old_object}{ANY. The "before" object.}

\item{new_object}{ANY. The "new" object. These are usually a data.frame or
an environment.}
}
\description{
Transform an atomic vector into an atomic vector.
}

