% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tracked_environment}
\alias{tracked_environment}
\title{An environment object that tracks changes made to its constituents.}
\usage{
tracked_environment(env)
}
\arguments{
\item{env}{environment. When converted to a \code{tracked_environment},
all changes will be remembered whenever a "commit" is registered
on the environment. Commits can be named and labeled.}
}
\description{
A typical R environment is like a mathematical set. It is a bag of
stuff and not much else. A tracked environment is like such an environment
that can also replay "changes" to its history.
}
\details{
To create a tracked environment from any environment \code{e}, simply write
\code{tracked_environment(e)}.
}
\examples{
\dontrun{
  e <- tracked_environment()
  e$x <- 1
  commit(e) <- 'First message'
  e$x <- 2
  commit(e) <- 'Second message'
  stopifnot(identical(e$x, 2))
  rollback(e) <- 1
  stopifnot(identical(e$x, 1)) # The changes have been rolled back one step.
}
}

